Reparaciones

+++ AccessControl/grails-app/controllers/org/votingsystem/accesscontrol/controller/CertificateVSController.groovy
@@ -8,9 +8,10 @@ import org.votingsystem.model.EnvironmentVS
 import org.votingsystem.util.ApplicationContextHolder;
 import org.votingsystem.model.ResponseVS
 import org.votingsystem.signature.util.CertUtil
-
+import org.votingsystem.util.HttpHelper
 import javax.xml.bind.annotation.adapters.HexBinaryAdapter
 import java.security.cert.X509Certificate
+
 /**
 * @infoController Servicio de Certificados
 * @descController Servicios relacionados con los certificates manejados por la aplicación
@@ -129,8 +130,8 @@ class CertificateVSController {
 	 */
 	def addCertificateAuthority () {
 		if(!EnvironmentVS.DEVELOPMENT.equals(ApplicationContextHolder.getEnvironment())) {
-            params.responseVS = new ResponseVS(ResponseVS.SC_ERROR_REQUEST,message(code: "serviceDevelopmentModeMsg"))
-			return
+            return params.responseVS = new ResponseVS(ResponseVS.SC_ERROR_REQUEST,
+                    message(code: "serviceDevelopmentModeMsg"))
 		}
 		log.debug "===============****¡¡¡¡¡ DEVELOPMENT Environment !!!!!****=================== "
 		signatureVSService.deleteTestCerts()
@@ -149,9 +150,49 @@ class CertificateVSController {
         params.responseVS = new ResponseVS(statusCode: ResponseVS.SC_OK, contentType: ContentTypeVS.PEM,
                 messageBytes: CertUtil.getPEMEncoded (trustedCerts))
 	}
-	
+
+    /**
+     * (SERVICIO DISPONIBLE SOLO EN ENTORNOS DE DESARROLLO). Servicio que borra los certificados generados en una
+     * simualcion
+     * @httpMethod [GET]
+     * @serviceURL [/certificateVS/deleteTestCerts]
+     * @return Los certificados en formato PEM de las Autoridades Certificadoras en las que
+     *         confía la aplicación.
+     */
 	def deleteTestCerts() {
+        if(!EnvironmentVS.DEVELOPMENT.equals(ApplicationContextHolder.getEnvironment())) {
+            return params.responseVS = new ResponseVS(ResponseVS.SC_ERROR_REQUEST,
+                    message(code: "serviceDevelopmentModeMsg"))
+        }
+        log.debug "===============****¡¡¡¡¡ DEVELOPMENT Environment !!!!!****=================== "
 		params.responseVS = signatureVSService.deleteTestCerts()
 	}
+
+    /**
+     * (SERVICIO DISPONIBLE SOLO EN ENTORNOS DE DESARROLLO). Servicio que da de alta el certificado del servidor
+     * en el servidor que recibe como parámetro.
+     * @httpMethod [GET]
+     * @serviceURL [/certificateVS/sendCA]
+     * @param [serverURL] Obligatorio. la URL base del servidor en el que se desea dar de alta el certificado.
+     */
+    def sendCA() {
+        if(!EnvironmentVS.DEVELOPMENT.equals(ApplicationContextHolder.getEnvironment())) {
+            return params.responseVS = new ResponseVS(ResponseVS.SC_ERROR_REQUEST,
+                    message(code: "serviceDevelopmentModeMsg"))
+        }
+        log.debug "===============****¡¡¡¡¡ DEVELOPMENT Environment !!!!!****=================== "
+        if(!params.serverURL) return params.responseVS = new ResponseVS(ResponseVS.SC_ERROR_REQUEST,
+                message(code: "missingParamErrorMsg", args:["serverURL"]))
+        else {
+            X509Certificate serverCert = signatureVSService.getServerCert()
+            byte[] rootCACertPEMBytes = CertUtil.getPEMEncoded (serverCert);
+            //String serviceURL = ActorVS.getRootCAServiceURL(params.serverURL)
+            String serviceURL = "${params.serverURL}/certificateVS/addCertificateAuthority"
+            return params.responseVS = HttpHelper.getInstance().sendData(rootCACertPEMBytes, ContentTypeVS.X509_CA,
+                    serviceURL);
+        }
+
+    }
+
 
 }
